requirements.txt
Django==1.11.5
jenkinsapi==0.3.4
mysql-connector-python==2.0.4
mysqlclient==1.3.7


